Optionally pass in ApiKey, instead of using AppSettings
In some cases (eg, Azure) AppSettings is not the preferred way to store
configuration information. To better handle those cases, added a new
StripeConfiguration class that can be passed in by users of the API. To
make that relatively painless, added a base class for all the *Service
classes, and made the Requestor non-static.